Add Voter

+++ src/Controller/FigureController.php
@@ -8,6 +8,7 @@ use App\Entity\Video;
 use App\Form\FigureType;
 use App\Repository\FigureRepository;
 use App\Repository\GroupRepository;
+use App\Security\Voter\FigureVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use App\Managers\FigureManager;
 use Exception;
@@ -46,7 +47,7 @@ class FigureController extends AbstractController
         PaginatorInterface $paginator
     ): Response
     {
-        $figures = $figureRepository->findAll();
+        $figures = $figureRepository->findBy([], ['created_at' => 'DESC']);
         $figures = $paginator->paginate(
             $figures,
             $request->query->getInt('page', 1),
@@ -75,6 +76,7 @@ class FigureController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $figure->setAuthor($this->getUser());
             $figureManager->save($figure);
             return $this->redirectToRoute('figure_index', [], Response::HTTP_SEE_OTHER);
         }
@@ -138,7 +140,6 @@ class FigureController extends AbstractController
      * Edit figure
      *
      * @Route("/figure/edit/{id}", name="figure_edit", methods={"GET", "POST"})
-     * @IsGranted("ROLE_ADMIN")
      * @param Request $request
      * @param FigureManager $figureManager
      * @param Figure $figure
@@ -152,9 +153,9 @@ class FigureController extends AbstractController
         GroupRepository $groupRepository
     ): Response
     {
+        $this->denyAccessUnlessGranted(FigureVoter::EDIT, $figure);
         $form = $this->createForm(FigureType::class, $figure);
         $form->handleRequest($request);
-
         if ($form->isSubmitted() && $form->isValid()) {
             $figureManager->save($figure);
             return $this->redirectToRoute('figure_index', [], Response::HTTP_SEE_OTHER);
@@ -171,7 +172,6 @@ class FigureController extends AbstractController
      * Delete figure
      *
      * @Route("/figure/delete/{id}", name="figure_delete", methods={"POST","GET"})
-     * @IsGranted("ROLE_ADMIN")
      * @param Request $request
      * @param Figure $figure
      * @param EntityManagerInterface $entityManager
@@ -179,6 +179,7 @@ class FigureController extends AbstractController
      */
     public function delete(Request $request, Figure $figure, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(FigureVoter::DELETE, $figure);
         if ($this->isCsrfTokenValid('delete' . $figure->getId(), $request->request->get('_token'))) {
             $entityManager->remove($figure);
             $entityManager->flush();
@@ -190,7 +191,6 @@ class FigureController extends AbstractController
      * Delete image from figure
      *
      * @Route("/figure/image/delete/{id}", name="figure_image_delete", methods={"POST","GET"})
-     * @IsGranted("ROLE_ADMIN")
      * @param Request $request
      * @param Image $image
      * @param EntityManagerInterface $entityManager
@@ -198,6 +198,7 @@ class FigureController extends AbstractController
      */
     public function deleteImage(Request $request, Image $image, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(FigureVoter::DELETE, $image->getFigure());
         if ($this->isCsrfTokenValid('delete-image' . $image->getId(), $request->request->get('_token'))) {
             $figure = $image->getFigure();
             $figure->removeImage($image);
@@ -211,7 +212,6 @@ class FigureController extends AbstractController
      * Edit image
      *
      * @Route("/figure/image/edit/{id}", name="figure_image_edit", methods={"POST","GET"})
-     * @IsGranted("ROLE_ADMIN")
      * @param Request $request
      * @param Image $image
      * @param EntityManagerInterface $entityManager
@@ -219,6 +219,7 @@ class FigureController extends AbstractController
      */
     public function updateImage(Request $request, Image $image, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(FigureVoter::EDIT, $image->getFigure());
         if ($this->isCsrfTokenValid('update-image' . $image->getId(), $request->request->get('_token'))) {
             $newImageUrl = $request->request->get('new_url');
             $image->setUrl($newImageUrl);
@@ -232,7 +233,6 @@ class FigureController extends AbstractController
      * Delete video
      *
      * @Route("/figure/video/delete/{id}", name="figure_video_delete", methods={"POST","GET"})
-     * @IsGranted("ROLE_ADMIN")
      * @param Request $request
      * @param Video $video
      * @param EntityManagerInterface $entityManager
@@ -240,6 +240,7 @@ class FigureController extends AbstractController
      */
     public function deleteVideo(Request $request, Video $video, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(FigureVoter::DELETE, $video->getFigure());
         if ($this->isCsrfTokenValid('delete-video' . $video->getId(), $request->request->get('_token'))) {
             $figure = $video->getFigure();
             $figure->removeVideo($video);
@@ -252,7 +253,6 @@ class FigureController extends AbstractController
      * Edit video
      *
      * @Route("/figure/video/edit/{id}", name="figure_video_edit", methods={"POST","GET"})
-     * @IsGranted("ROLE_ADMIN")
      * @param Request $request
      * @param Video $video
      * @param EntityManagerInterface $entityManager
@@ -260,6 +260,7 @@ class FigureController extends AbstractController
      */
     public function editVideo(Request $request, Video $video, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(FigureVoter::EDIT, $video->getFigure());
         if ($this->isCsrfTokenValid('update-video' . $video->getId(), $request->request->get('_token'))) {
             $newVideoUrl = $request->request->get('new_url');
             $video->setUrl($newVideoUrl);
